+++ src/components/PluginLoader.jsx
@@ -1,9 +1,9 @@
-import { useAlert, useConfig } from '@dhis2/app-runtime'
 // eslint-disable-next-line import/no-unresolved
 import { Plugin } from '@dhis2/app-runtime/experimental'
+import { CircularLoader, CenteredContent, NoticeBox } from '@dhis2/ui'
 import PropTypes from 'prop-types'
-import React from 'react'
-import { useLocation, useParams } from 'react-router-dom'
+import React, { useCallback, useMemo, useState } from 'react'
+import { Link, useLocation, useParams } from 'react-router'
 import { useClientOfflineInterface } from '../lib/clientPWAUpdateState.jsx'
 import i18n from '../locales/index.js'
 import styles from './PluginLoader.module.css'
@@ -32,30 +32,11 @@ const injectHeaderbarHidingStyles = (event) => {
 }
 
 // todo: this is kinda duplicated between here and the header bar
-const getAppDefaultAction = (appName, modules) => {
+const getPluginEntrypoint = (appName, modules) => {
     // If core apps get a different naming scheme, this needs revisiting
     return modules.find(
         (m) => m.name === appName || m.name === 'dhis-web-' + appName
     )?.defaultAction
-}
-
-// todo: why is there a request to /apps/undefined?redirect=false
-const getPluginEntrypoint = async (appName, modules /* baseUrl */) => {
-    const defaultAction = getAppDefaultAction(appName, modules)
-
-    // todo: app.html handling ----
-    // todo: this could be better if this can be detected from app entrypoints API
-    // const defaultAppUrl = new URL(defaultAction, baseUrl)
-    // const pluginifiedAppUrl = new URL('./app.html', defaultAppUrl)
-
-    // Start by trying to load pluginified app, `app.html`
-    // const pluginifiedAppResponse = await fetch(pluginifiedAppUrl)
-    // if (pluginifiedAppResponse.ok) {
-    //     return pluginifiedAppUrl.href
-    // }
-
-    // If pluginified app is not found, fall back to app root
-    return defaultAction
 }
 
 const watchForHashRouteChanges = (event) => {
@@ -78,6 +59,8 @@ const listenForCommandPaletteToggle = (event) => {
 }
 
 /**
+ * ⭐️ This is what redirects back to the regular login app when logged out ⭐️
+ *
  * If the iframe loads a page that is different from the pluginSource given to
  * it, navigate the whole page there. This should handle two cases:
  * 1. The navigation is outside the DHIS2 instance: we want to leave the shell
@@ -100,71 +83,57 @@ const handleExternalNavigation = (iframeLoadEvent, pluginHref) => {
     }
 }
 
+const failedLoadErrorMessage =
+    'The requested page is not accessible by the DHIS2 global shell, ' +
+    'and the URL is therefore inaccessible to be printed here. ' +
+    'The server might be down, or the requested page is on a ' +
+    'different domain. ' +
+    'In the second case, the link should be opened in a new tab instead.'
+
 export const PluginLoader = ({ appsInfoQuery }) => {
     const params = useParams()
     const location = useLocation()
-    const { baseUrl } = useConfig()
-    const [pluginEntrypoint, setPluginEntrypoint] = React.useState()
-    const [rerenderKey, setRerenderKey] = React.useState(0)
-    const { show: showNavigationWarning } = useAlert(
-        i18n.t(
-            'Unable to load the requested page from DHIS2. Returned to previous page.'
-        ),
-        { warning: true }
-    )
     const initClientOfflineInterface = useClientOfflineInterface()
+    const [error, setError] = useState(null)
 
-    // test prop messaging and updates
-    const [color, setColor] = React.useState('blue')
-    const toggleColor = React.useCallback(
-        () => setColor((prev) => (prev === 'blue' ? 'red' : 'blue')),
-        []
-    )
-
-    // todo: add query string to entrypoint (e.g. maps /dhis-web-maps/?currentAnalyticalObject=true)
-    // Can use `urlObject.searchParams.append('redirect', 'false')`
-    // (or is this a backend thing?)
-    React.useEffect(() => {
+    const pluginHref = useMemo(() => {
         if (!appsInfoQuery.data) {
             return
         }
-        // Performed async to test for index/app entrypoint
-        // todo: this could be better if this can be detected from app entrypoints API
-        const asyncWork = async () => {
-            // for testing: params.appName === 'localApp' ? 'http://localhost:3001/app.html'
-            const newPluginEntrypoint = await getPluginEntrypoint(
-                params.appName,
-                appsInfoQuery.data.appMenu.modules,
-                baseUrl
-            )
-            setPluginEntrypoint(newPluginEntrypoint)
-        }
-        asyncWork()
-    }, [params.appName, baseUrl, appsInfoQuery.data])
+        setError(null)
 
-    const pluginHref = React.useMemo(() => {
+        // for testing: params.appName === 'localApp' ? 'http://localhost:3001/app.html'
+        const newPluginEntrypoint = getPluginEntrypoint(
+            params.appName,
+            appsInfoQuery.data.appMenu.modules
+        )
+
+        if (!newPluginEntrypoint) {
+            console.error(
+                `The app slug "${params.appName}" did not match any app.`
+            )
+            setError(i18n.t('Unable to find an app for this URL.'))
+            return
+        }
+
         // An absolute URL helps compare to the location inside the iframe:
-        const pluginUrl = new URL(pluginEntrypoint, window.location)
+        const pluginUrl = new URL(newPluginEntrypoint, window.location)
         pluginUrl.hash = location.hash
         pluginUrl.search = location.search
         pluginUrl.searchParams.append('redirect', 'false')
-        return pluginUrl.href
-    }, [pluginEntrypoint, location.hash, location.search])
 
-    const handleLoad = React.useCallback(
+        return pluginUrl.href
+    }, [location.hash, location.search, appsInfoQuery.data, params.appName])
+
+    const handleLoad = useCallback(
         (event) => {
             // If we can't access the new page's Document, this is a cross-domain page.
-            // Disallow that; return to previous plugin state.
-            // todo: can cause an infinite reload if the current pluginHref loads
-            // an entirely broken page -- this is a rare case though; one example is
-            // a PWA app where the precache has been deleted. 404s are fine.
+            // Disallow that and show an error
             if (!event.target.contentDocument) {
-                setRerenderKey((k) => k + 1)
-                showNavigationWarning()
-                console.error(
-                    'The linked page is not accessible by the DHIS2 global shell; returned to previous plugin state. ' +
-                        'This link should be opened in a new tab instead'
+                setError(
+                    i18n.t('The requested page is not accessible from DHIS2.')
                 )
+                console.error(failedLoadErrorMessage)
                 return
             }
 
@@ -178,11 +147,28 @@ export const PluginLoader = ({ appsInfoQuery }) => {
             })
             listenForCommandPaletteToggle(event)
         },
-        [pluginHref, showNavigationWarning, initClientOfflineInterface]
+        [pluginHref, initClientOfflineInterface]
     )
 
+    if (error) {
+        return (
+            <CenteredContent>
+                <NoticeBox title={i18n.t('Something went wrong')} error>
+                    <div className={styles.marginBottom}>{error}</div>
+                    <Link to={'/'} className={styles.link}>
+                        {i18n.t('Return to home page')}
+                    </Link>
+                </NoticeBox>
+            </CenteredContent>
+        )
+    }
+
     if (!pluginHref) {
-        return 'Loading...' // todo
+        return (
+            <CenteredContent>
+                <CircularLoader />
+            </CenteredContent>
+        )
     }
 
     return (
@@ -191,11 +177,6 @@ export const PluginLoader = ({ appsInfoQuery }) => {
             // pass URL hash down to the client app
             pluginSource={pluginHref}
             onLoad={handleLoad}
-            key={rerenderKey}
-            // Other props
-            // (for testing:)
-            color={color}
-            toggleColor={toggleColor}
         />
     )
 }