+++ src/components/header-bar/command-palette/views/home-view.jsx
@@ -6,7 +6,13 @@ import AppItem from '../sections/app-item.jsx'
 import Heading from '../sections/heading.jsx'
 import ListItem from '../sections/list-item.jsx'
 
-const HomeView = ({ grid, gridColumnCount, gridRowCount, currentItem }) => {
+const HomeView = ({
+    grid,
+    gridColumnCount,
+    gridRowCount,
+    currentItem,
+    resetModal,
+}) => {
     return (
         <>
             {gridRowCount > 0 && (
@@ -27,10 +33,12 @@ const HomeView = ({ grid, gridColumnCount, gridRowCount, currentItem }) => {
                                 return (
                                     <AppItem
                                         key={`app-${name}-${idx}`}
-                                        name={displayName || name}
+                                        name={name}
+                                        displayName={displayName || name}
                                         path={defaultAction}
                                         img={icon}
                                         highlighted={item === currentItem}
+                                        resetModal={resetModal}
                                     />
                                 )
                             })
@@ -91,6 +99,7 @@ HomeView.propTypes = {
     grid: PropTypes.array,
     gridColumnCount: PropTypes.number,
     gridRowCount: PropTypes.number,
+    resetModal: PropTypes.func,
 }
 
 export default HomeView