+++ src/components/header-bar/command-palette/views/list-view.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import EmptySearchResults from '../sections/empty-search-results.jsx'
 import ListItem from '../sections/list-item.jsx'
 
-const ListView = ({ grid, currentItem }) => {
+const ListView = ({ grid, currentItem, resetModal }) => {
     const listItems = grid.reduce((acc, arr) => {
         acc.push(arr[0])
         return acc
@@ -18,10 +18,8 @@ const ListView = ({ grid, currentItem }) => {
                             action,
                             displayName,
                             name,
-                            defaultAction,
                             icon,
                             description,
-                            url,
                             type,
                             dataTest,
                         } = item
@@ -31,15 +29,16 @@ const ListView = ({ grid, currentItem }) => {
                         return (
                             <ListItem
                                 type={type}
-                                key={`app-${name}-${idx}`}
+                                key={`list-item-${idx}-${name}`}
+                                name={name}
                                 title={displayName || name}
-                                path={defaultAction || url}
                                 image={isImage ? icon : undefined}
                                 icon={isIcon ? icon : undefined}
                                 description={description}
                                 highlighted={currentItem === item}
                                 onClickHandler={action}
                                 dataTest={dataTest}
+                                resetModal={resetModal}
                             />
                         )
                     })}
@@ -54,6 +53,7 @@ const ListView = ({ grid, currentItem }) => {
 ListView.propTypes = {
     currentItem: PropTypes.object,
     grid: PropTypes.array,
+    resetModal: PropTypes.func,
 }
 
 export default ListView