+++ src/components/header-bar/profile.jsx
@@ -1,26 +1,23 @@
 import { spacers } from '@dhis2/ui-constants'
+import { Layer } from '@dhis2-ui/layer'
+import { Popper } from '@dhis2-ui/popper'
 import { UserAvatar } from '@dhis2-ui/user-avatar'
 import PropTypes from 'prop-types'
 import React, { useCallback, useRef, useState } from 'react'
 import i18n from '../../locales/index.js'
 import { DebugInfoModal } from './debug-info/debug-info-modal.jsx'
-import { useOnDocClick } from './profile/use-on-doc-click.js'
 import { ProfileMenu } from './profile-menu/index.js'
 
 const Profile = ({ name, avatarId, helpUrl, username }) => {
     const [showProfileMenu, setShowProfileMenu] = useState(false)
     const [showDebugInfoModal, setShowDebugInfoModal] = useState(false)
-    const hideProfileMenu = useCallback(
-        () => setShowProfileMenu(false),
-        [setShowProfileMenu]
-    )
+
+    const containerRef = useRef(null)
+    const hideProfileMenu = useCallback(() => setShowProfileMenu(false), [])
     const toggleProfileMenu = useCallback(
         () => setShowProfileMenu((show) => !show),
-        [setShowProfileMenu]
+        []
     )
-    const containerRef = useRef(null)
-
-    useOnDocClick(containerRef, hideProfileMenu)
 
     return (
         <div
@@ -43,16 +40,20 @@ const Profile = ({ name, avatarId, helpUrl, username }) => {
             </button>
 
             {showProfileMenu && (
-                <ProfileMenu
-                    avatarId={avatarId}
-                    name={name}
-                    username={username}
-                    helpUrl={helpUrl}
-                    hideProfileMenu={hideProfileMenu}
-                    showDebugInfoModal={() => {
-                        setShowDebugInfoModal(true)
-                    }}
-                />
+                <Layer onBackdropClick={hideProfileMenu}>
+                    <Popper reference={containerRef} placement="bottom-end">
+                        <ProfileMenu
+                            avatarId={avatarId}
+                            name={name}
+                            username={username}
+                            helpUrl={helpUrl}
+                            hideProfileMenu={hideProfileMenu}
+                            showDebugInfoModal={() => {
+                                setShowDebugInfoModal(true)
+                            }}
+                        />
+                    </Popper>
+                </Layer>
             )}
             {showDebugInfoModal && (
                 <DebugInfoModal